pyproject.toml
[project]
name = "autodev"
version = "0.1.0"
description = "Personal environment for agent-assisted development using a throwaway virtual machine"
requires-python = ">=3.8"
dependencies = [
    "hcloud>=1.20.0",
    "python-telegram-bot>=21.0.0",
    "python-decouple>=3.8",
    "python-dotenv>=1.0.0",
    "configargparse>=1.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
